remove FAILED_CHECK constants, move util.ts into checks folder
