+++ packages/step-checker/src/checks/__tests__/fraction-checks.test.ts
@@ -2,7 +2,7 @@ import {serializer} from "@math-blocks/semantic";
 import {parse} from "@math-blocks/text-parser";
 
 import StepChecker from "../../step-checker";
-import {deepEquals} from "../../util";
+import {deepEquals} from "../util";
 import {checkStep} from "../test-util";
 
 expect.addSnapshotSerializer(serializer);
