+++ packages/step-checker/src/checks/axiom-checks.ts
@@ -1,23 +1,18 @@
 import * as Semantic from "@math-blocks/semantic";
 
-import {zip, applySteps} from "../util";
+import {zip, applySteps} from "./util";
 import {Result, Step, Check} from "../types";
-import {FAILED_CHECK} from "../constants";
 
 import {exactMatch, checkArgs} from "./basic-checks";
 
 export const addZero: Check = (prev, next, context) => {
-    return prev.type === "add"
-        ? checkIdentity(prev, next, context)
-        : FAILED_CHECK;
+    return prev.type === "add" ? checkIdentity(prev, next, context) : undefined;
 };
 
 addZero.symmetric = true;
 
 export const mulOne: Check = (prev, next, context) => {
-    return prev.type === "mul"
-        ? checkIdentity(prev, next, context)
-        : FAILED_CHECK;
+    return prev.type === "mul" ? checkIdentity(prev, next, context) : undefined;
 };
 
 mulOne.symmetric = true;
@@ -43,7 +38,7 @@ export const checkIdentity: Check<Semantic.Add | Semantic.Mul> = (
 
     // If we haven't removed any identities then this check has failed
     if (nonIdentityArgs.length === prev.args.length) {
-        return FAILED_CHECK;
+        return;
     }
 
     // Steps are local to the nodes involved which are descendents of prev so
@@ -92,7 +87,7 @@ export const checkIdentity: Check<Semantic.Add | Semantic.Mul> = (
         };
     }
 
-    return FAILED_CHECK;
+    return;
 };
 
 checkIdentity.symmetric = true;
@@ -201,7 +196,7 @@ export const checkDistribution: Check = (prev, next, context) => {
     }
 
     if (prev.type !== "mul" || next.type !== "add") {
-        return FAILED_CHECK;
+        return;
     }
 
     // If the second factor is an add, e.g. a(b + c) -> ...
@@ -268,14 +263,14 @@ export const checkDistribution: Check = (prev, next, context) => {
             };
         }
     }
-    return FAILED_CHECK;
+    return;
 };
 
 checkDistribution.symmetric = true;
 
 export const mulByZero: Check = (prev, next, context) => {
     if (prev.type !== "mul") {
-        return FAILED_CHECK;
+        return;
     }
 
     // TODO: ensure that steps from these calls to checkStep
@@ -299,7 +294,7 @@ export const mulByZero: Check = (prev, next, context) => {
             ],
         };
     }
-    return FAILED_CHECK;
+    return;
 };
 
 mulByZero.symmetric = true;
@@ -317,7 +312,7 @@ export const commuteAddition: Check = (prev, next, context) => {
 
         // If they aren't we can stop this check right here.
         if (!result) {
-            return FAILED_CHECK;
+            return;
         }
 
         // If at least some of the pairs don't line up then it's safe to
@@ -361,7 +356,7 @@ export const commuteAddition: Check = (prev, next, context) => {
         }
     }
 
-    return FAILED_CHECK;
+    return;
 };
 
 export const commuteMultiplication: Check = (prev, next, context) => {
@@ -377,7 +372,7 @@ export const commuteMultiplication: Check = (prev, next, context) => {
 
         // If the args are the same then we can stop here.
         if (!result) {
-            return FAILED_CHECK;
+            return;
         }
 
         const reordered = pairs.some(
@@ -410,7 +405,7 @@ export const commuteMultiplication: Check = (prev, next, context) => {
         }
     }
 
-    return FAILED_CHECK;
+    return;
 };
 
 export const symmetricProperty: Check = (prev, next, context) => {
@@ -419,7 +414,7 @@ export const symmetricProperty: Check = (prev, next, context) => {
     // and then swap sides at the last moment so that the variable that we're
     // looking to isolate is on the left.
     if (!context.reversed) {
-        return FAILED_CHECK;
+        return;
     }
 
     if (
@@ -480,7 +475,7 @@ export const symmetricProperty: Check = (prev, next, context) => {
         }
     }
 
-    return FAILED_CHECK;
+    return;
 };
 
 symmetricProperty.symmetric = true;
