+++ packages/step-checker/src/checks/polynomial-checks.ts
@@ -1,10 +1,9 @@
 import {Check} from "../types";
-import {FAILED_CHECK} from "../constants";
 
 // TODO: Implement this.
 // It should handle things like: 2a + 3 + 5a + 7 -> 7a + 10
 const collectLikeTerms: Check = (prev, next, context) => {
-    return FAILED_CHECK;
+    return undefined;
 };
 
 export const runChecks: Check = (prev, next, context) => {
@@ -13,5 +12,5 @@ export const runChecks: Check = (prev, next, context) => {
         return result;
     }
 
-    return FAILED_CHECK;
+    return;
 };
