--- packages/step-checker/src/constants.ts
@@ -1,1 +1,0 @@
-export const FAILED_CHECK = undefined;
